Move some classes to optional modules ('com.intellij.modules.vcs') to add support for 'Code With Me'
