+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 - \#309 Exception in plugin Jenkins Control
 - \#322 Nullpointer if Jenkins if not running
+- move classes with module 'com.intellij.modules.vcs' to optional dependency
 
 ## [0.13.12]
 - \#280 render DateParameterDefinition as textfield
