+++ src/main/resources/META-INF/org.codinjutsu.tools.jenkins-JobParameterRenderer.xml
@@ -1,5 +1,4 @@
 <idea-plugin>
-
     <extensions defaultExtensionNs="Jenkins Control Plugin">
         <buildParameterRenderer implementation="org.codinjutsu.tools.jenkins.view.parameter.BuiltInJobParameterRenderer"
                                 order="first"
