+++ src/main/resources/META-INF/org.codinjutsu.tools.jenkins-LangSupport.xml
@@ -1,0 +1,2 @@
+<idea-plugin>
+</idea-plugin>
